# api/swagger/swagger.yaml
swagger: "2.0"
info:
  title: Vitality
  version: v1
#host: my.remote.host
host: localhost:3000
basePath: /
schemes:
  - http
  - https
consumes:
- application/json
- application/javascript
produces:
- application/json
- application/javascript
- text/javascript
paths:
  /loadallrepos:
    x-swagger-router-controller: core_loadallrepos
    get:
      summary: Initialize the application with repository data
      description: |
        Initializes the application by loading information for select repositories into the application state; forks are filtered out. 
      operationId: loadallrepos
      parameters:
        - name: callback
          in: query
          description: Passed by .ajax call for cross-domain requests
          required: false
          type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/loadallreposResponse"
        400:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  ErrorResponse:
    type: string
    required:
      - success
      - message
    properties:
      success:
        type: string
      message:
        type: string
      nativemessage:
        type: string
      nativecode:
        type: integer
...
  loadallreposResponse:
    type: string
    required:
      - success
      - data
    properties:
      success:
        type: boolean
      data:
        type: object
        required:
          - nextrefresh
        properties:
          nextrefresh:
            type: integer
...